+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fix race condition between ready and task functions worker message handling at startup.
 - Fix duplicate task usage statistics computation per task function.
+- Update task function worker usage statistics if and only if there's at least two different task functions.
 
 ### Added